chore: Refactor logpipeline validator
